ai-endpoints/depth-estimation/requirements.txt
torch
opencv-python
firebase-admin
numpy
python-dotenv
timm